+++ modyn/common/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_subdirectory("cpp")
 add_subdirectory("trigger_sample")
 add_subdirectory("example_extension")
-add_subdirectory("local_storage_backend")
