+++ modyn/selector/internal/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory("storage_backend")
