+++ modyn/selector/internal/storage_backend/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory("local")
